+++ src/chart/rendering/drawdiagram.cpp
@@ -25,6 +25,7 @@ drawDiagram::drawDiagram(const Geom::Rect &rect,
 		drawBackground(rect, canvas, style.plot, events.plot.background,
 			Events::OnRectDrawParam("plot"));
 
+		drawArea(false);
 		drawAxes(*this).drawBase();
 	}
 
@@ -52,7 +53,11 @@ drawDiagram::drawDiagram(const Geom::Rect &rect,
 void drawDiagram::clipPlotArea()
 {
 	canvas.save();
-	
+	drawArea(true);
+}
+
+void drawDiagram::drawArea(bool clip)
+{
 	std::array<Geom::Point, 4> points =
 	{
 		Geom::Point(0.0, 0.0), 
@@ -63,7 +68,23 @@ void drawDiagram::clipPlotArea()
 	painter.setPolygonToCircleFactor(0.0);
 	painter.setPolygonStraightFactor(0.0);
 	painter.setResMode(ResolutionMode::High);
-	painter.drawPolygon(points, true);
+
+	if (clip) 
+	{
+		painter.drawPolygon(points, true);
+	}
+	else
+	{
+
+		if (!style.plot.areaColor->isTransparent())
+		{
+			canvas.setBrushColor(*style.plot.areaColor);
+			canvas.setLineColor(*style.plot.areaColor);
+			canvas.setLineWidth(0);
+				painter.drawPolygon(points, false);
+			canvas.setLineWidth(0);
+		}
+	}
 }
 
 void drawDiagram::drawMarkerGuides()